requirements.txt
PyYAML==6.0.2
gradio==5.1.0
displaytree==1.1.0
langchain==0.3.4
langchain-community==0.3.3
langchain-openai==0.2.3
directory-tree==1.0.0
chromadb==0.5.0 
langchain-anthropic==0.2.3
beautifulsoup4==4.12.3
langgraph==0.2.39
datasets==3.0.2